move refresh button
