+++ webapp/javascript/ui/Button.module.scss
@@ -21,13 +21,18 @@ $dark-text: #222;
   border-radius: $border-radius;
 
   /* when we have an icon only button */
-  min-width: 48px;
+  //min-width: 48px;
 }
 
 .icon {
   // icon is on the left
   // some space between the icon and text
   margin-right: 5px;
+}
+// undo the margin right
+// for when the icon is alone
+.icon:only-child {
+  margin-right: 0;
 }
 
 .default {
